Balance the game to be a bit less unforgiving
* Generally reduced monster hit points by ca 25-30%
* Slightly reduced shock taken from seeing monsters
* Reduced the standard duration of many negative properties
  (e.g. confusion from gas traps last 30-60 turns instead of 80-120
  turns)
* Removed the hit point penalty for the Occultist class

To compensate for the monster hit point change, the following changes
were made:

* The Rogue background starts with a normal dagger instead of +1
* Reduced dagger backstab damage bonus from +300% to +200%
* Reduced vicious/ruthless backstab damage bonus from +150% to +100%

+++ src/game.cpp
@@ -130,7 +130,7 @@ void save()
         saving::put_int(s_start_time.minute);
         saving::put_int(s_start_time.second);
 
-        saving::put_int(s_history_events.size());
+        saving::put_int((int)s_history_events.size());
 
         for (const HistoryEvent& event : s_history_events)
         {
